+++ Listener.php
@@ -4,15 +4,22 @@ namespace TickTackk\UpdateAnyResource;
 
 use XF\Service\User\ContentChange as UserContentChangeSvc;
 
+/**
+ * @version 1.1.1
+ */
 class Listener
 {
     /**
      * @param UserContentChangeSvc $changeService
-     * @param array                updates
+     * @param array $updates
+     *
+     * @return void
+     *
+     * @noinspection PhpUnusedParameterInspection
      */
-    public static function userContentChangeInit(UserContentChangeSvc $changeService, array &$updates)
+    public static function userContentChangeInit(UserContentChangeSvc $changeService, array &$updates) : void
     {
-        $updates['xf_rm_resource_update'] = ['user_id', 'username'];
-        $updates['xf_rm_resource_version'] = ['user_id', 'username'];
+        $updates['xf_rm_resource_update'] = ['tck_uar_user_id', 'tck_uar_username'];
+        $updates['xf_rm_resource_version'] = ['tck_uar_user_id', 'tck_uar_username'];
     }
 }