+++ XFRM/Entity/ResourceUpdate.php
@@ -2,19 +2,20 @@
 
 namespace TickTackk\UpdateAnyResource\XFRM\Entity;
 
-use TickTackk\UpdateAnyResource\Globals;
+use TickTackk\UpdateAnyResource\XFRM\Entity\XF21\ResourceItem as XF21ResourceItemEntity;
+use TickTackk\UpdateAnyResource\XFRM\Entity\XF22\ResourceItem as XF22ResourceItemEntity;
 use XF\Mvc\Entity\Structure;
 use XF\Phrase;
 use XF\Entity\User as UserEntity;
 
 /**
  * COLUMNS
- * @property int user_id
- * @property int user_id_
- * @property string username
+ * @property int $tck_uar_user_id
+ * @property string $tck_uar_username
  *
  * RELATIONS
- * @property UserEntity User
+ * @property UserEntity $User
+ * @property ResourceItemTrait|XF21ResourceItemEntity|XF22ResourceItemEntity $Resource
  */
 class ResourceUpdate extends XFCP_ResourceUpdate
 {
@@ -25,15 +26,16 @@ class ResourceUpdate extends XFCP_ResourceUpdate
      */
     public function canEdit(&$error = null)
     {
-        try
+        $resource = $this->Resource;
+        if (!$resource)
         {
-            Globals::$makeUseOfUpdateAnyPermission = $this;
+            return false;
+        }
+
+        return $resource->runActionForTckUpdateAnyResource($this->tck_uar_user_id, function () use($error)
+        {
             return parent::canEdit($error);
-        }
-        finally
-        {
-            Globals::$makeUseOfUpdateAnyPermission = null;
-        }
+        });
     }
 
     /**
@@ -44,28 +46,16 @@ class ResourceUpdate extends XFCP_ResourceUpdate
      */
     public function canDelete($type = 'soft', &$error = null)
     {
-        $visitor = \XF::visitor();
         $resource = $this->Resource;
-
-        if (!$visitor->user_id|| !$resource || $this->isDescription())
+        if (!$resource)
         {
             return false;
         }
 
-        if ($type !== 'soft')
+        return $resource->runActionForTckUpdateAnyResource($this->tck_uar_user_id, function () use($type, $error)
         {
-            return $resource->hasPermission('hardDeleteAny');
-        }
-
-        if ($resource->hasPermission('deleteAny'))
-        {
-            return true;
-        }
-
-        return (
-            $this->user_id === $visitor->user_id
-            && $resource->hasPermission('updateOwn')
-        );
+            return parent::canDelete($type, $error);
+        });
     }
 
     /**
@@ -73,14 +63,9 @@ class ResourceUpdate extends XFCP_ResourceUpdate
      */
     public function canSendModeratorActionAlert()
     {
-        $visitor = \XF::visitor();
-        $resource = $this->Resource;
-
         return (
-            $resource
-            && $resource->canSendModeratorActionAlert()
-            && $this->message_state === 'visible'
-            && $this->user_id !== $visitor->user_id
+            parent::canSendModeratorActionAlert()
+            && (\XF::visitor()->user_id !== $this->tck_uar_user_id)
         );
     }
 
@@ -93,14 +78,16 @@ class ResourceUpdate extends XFCP_ResourceUpdate
     {
         $structure = parent::getStructure($structure);
 
-        $structure->columns['user_id'] = ['type' => self::UINT, 'required' => true];
-        $structure->columns['username'] = ['type' => self::STR, 'maxLength' => 50,
+        $structure->columns['tck_uar_user_id'] = ['type' => self::UINT, 'required' => true];
+        $structure->columns['tck_uar_username'] = ['type' => self::STR, 'maxLength' => 50,
             'required' => 'please_enter_valid_name'
         ];
         $structure->relations['User'] = [
             'entity' => 'XF:User',
             'type' => self::TO_ONE,
-            'conditions' => 'user_id',
+            'conditions' => [
+                ['user_id', '=', '$tck_uar_user_id']
+            ],
             'primary' => true
         ];