+++ _output/extension_hint.php
@@ -17,17 +17,11 @@ namespace TickTackk\UpdateAnyResource\XFRM\Notifier\ResourceUpdate
 
 namespace TickTackk\UpdateAnyResource\XFRM\Service\ResourceItem
 {
-	class XFCP_ChangeType extends \XFRM\Service\ResourceItem\ChangeType {}
-	class XFCP_Create extends \XFRM\Service\ResourceItem\Create {}
 	class XFCP_Reassign extends \XFRM\Service\ResourceItem\Reassign {}
 }
 
 namespace TickTackk\UpdateAnyResource\XFRM\Service\ResourceUpdate
 {
 	class XFCP_Create extends \XFRM\Service\ResourceUpdate\Create {}
-}
-
-namespace TickTackk\UpdateAnyResource\XFRM\Service\ResourceVersion
-{
-	class XFCP_Create extends \XFRM\Service\ResourceVersion\Create {}
+	class XFCP_Preparer extends \XFRM\Service\ResourceUpdate\Preparer {}
 }