+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 CHANGELOG
 ==========================
 
-## 1.1.0 (`1000170`)
+## 1.1.1 (`1000170`)
+
+- **New:** Allow the ability to replicate native create thread reply for resource update under resource owner (#10)
+- **Fix:** Since 2.2, moderator action alert cannot be sent if the user is a team member (#6)
+- **Fix:** Since 2.2, team members are unable to post new resource updates (#5)
+- **Fix:** User who is not team member and made the release cannot download without permission (#7)
+- **Fix:** Username/user id columns added in resource update/version are not prefixed (#8)
+- **Fix:** Since 2.2, Resource update thread reply is made by resource owner (#9)
+- **Fix:** User is alerted for new resource update when ignoring the resource owner (#11)
+
+## 1.1.0 (`1000070`)
 
 - **Change:** Increase minimum XenForo version requirement to 2.1.10 Patch 2
 - **Change:** Increase minimum XenForo Resource Manager version requirement to 2.1.4